FLOAT type support (#143)

+++ test/sql/prepared/test_prepared.cpp
@@ -1,7 +1,7 @@
 #include "catch.hpp"
 #include "common/file_system.hpp"
-#include "test_helpers.hpp"
 #include "common/types/date.hpp"
+#include "test_helpers.hpp"
 
 using namespace duckdb;
 using namespace std;
@@ -127,7 +127,8 @@ TEST_CASE("PREPARE many types for INSERT", "[prepared]") {
 	Connection con(db);
 
 	// prepare different types in insert
-	REQUIRE_NO_FAIL(con.Query("CREATE TABLE test(a TINYINT, b SMALLINT, c INTEGER, d BIGINT, e REAL, f DOUBLE, g DATE, h VARCHAR)"));
+	REQUIRE_NO_FAIL(con.Query(
+	    "CREATE TABLE test(a TINYINT, b SMALLINT, c INTEGER, d BIGINT, e REAL, f DOUBLE, g DATE, h VARCHAR)"));
 	REQUIRE_NO_FAIL(con.Query("PREPARE s1 AS INSERT INTO test VALUES ($1,$2,$3,$4,$5,$6,$7,$8);"));
 	REQUIRE_NO_FAIL(con.Query("EXECUTE s1(1,2,3,4,1.5,2.5,'1992-10-20', 'hello world');"));
 	result = con.Query("SELECT * FROM test");
@@ -135,7 +136,7 @@ TEST_CASE("PREPARE many types for INSERT", "[prepared]") {
 	REQUIRE(CHECK_COLUMN(result, 1, {2}));
 	REQUIRE(CHECK_COLUMN(result, 2, {3}));
 	REQUIRE(CHECK_COLUMN(result, 3, {4}));
-	REQUIRE(CHECK_COLUMN(result, 4, {1.5}));
+	REQUIRE(CHECK_COLUMN(result, 4, {(float)1.5}));
 	REQUIRE(CHECK_COLUMN(result, 5, {2.5}));
 	REQUIRE(CHECK_COLUMN(result, 6, {Value::DATE(1992, 10, 20)}));
 	REQUIRE(CHECK_COLUMN(result, 7, {"hello world"}));
